refactor(speech): moved wav functionality to external class
